fix sentry reporter
Sometimes we cannot access sys.exc_info(). TurboGears saves the original
exc_info() in the os.environ so that we can properly initialize the
Traceback object. I'm taking advantage of this information to specify
the exc_info attribute of the method captureException from the raven
lib. This way it won't need to access sys.exc_info().